efi: Break down mod.rs into multiple files
